requirements.txt
aiogram
aiohttp
motor
pymongo
telethon
pyrogram
requests
pysocks
aioredis
envparse
pyyaml
redis
translate-api
sqlalchemy
pymongo[srv]
python-telegram-bot
envparse
rejson
numpy
googletrans==3.1.0a0